+++ src/libkeymap/keymap/array.h
@@ -1,6 +1,8 @@
 #ifndef LK_ARRAY_H
 #define LK_ARRAY_H
 
+#include <sys/param.h>
+
 /**
  * @brief Basic structure for array implementation.
  * @details The array is designed to store an arbitrary number of similar items.
@@ -19,11 +21,11 @@ int lk_array_empty(struct lk_array *a);
 
 int lk_array_append(struct lk_array *a, const void *e);
 
-int lk_array_set(struct lk_array *a, unsigned int i, const void *e);
-void *lk_array_get(struct lk_array *a, unsigned int i);
-void *lk_array_get_ptr(struct lk_array *a, unsigned int i);
+int lk_array_set(struct lk_array *a, size_t i, const void *e);
+void *lk_array_get(struct lk_array *a, size_t i);
+void *lk_array_get_ptr(struct lk_array *a, size_t i);
 
-int lk_array_unset(struct lk_array *a, unsigned int i);
-int lk_array_exists(struct lk_array *a, unsigned int i);
+int lk_array_unset(struct lk_array *a, size_t i);
+int lk_array_exists(struct lk_array *a, size_t i);
 
 #endif /* LK_ARRAY_H */
